refactor set_backend to common and use in other modules as well
