+++ analyze_screen_recording.py
@@ -7,20 +7,7 @@ import torch
 import torchvision.transforms as transforms
 from PIL import Image
 
-
-def set_backend():
-    use_mps = torch.backends.mps.is_available()
-    print(f"MPS available: {use_mps}")
-
-    # If MPS is available, use it. Otherwise, use CUDA if available, else use CPU
-    if use_mps:
-        device = torch.device("mps")
-    elif torch.cuda.is_available():
-        device = torch.device("cuda")
-    else:
-        device = torch.device("cpu")
-    print(f"Using device: {device}")
-    return device
+from common import set_backend
 
 
 def load_checkpoint(checkpoint_path, device):
