+++ common.py
@@ -1,0 +1,16 @@
+import torch
+
+
+def set_backend():
+    use_mps = torch.backends.mps.is_available()
+    print(f"MPS available: {use_mps}")
+
+    # If MPS is available, use it. Otherwise, use CUDA if available, else use CPU
+    if use_mps:
+        device = torch.device("mps")
+    elif torch.cuda.is_available():
+        device = torch.device("cuda")
+    else:
+        device = torch.device("cpu")
+    print(f"Using device: {device}")
+    return device
